requirements.txt
# Core Dependencies
streamlit==1.29.0
Pillow==10.1.0
opencv-python==4.8.1

# LangChain & LLM
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6

# RAG & Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
numpy==1.24.3